requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
dask==2024.3.1
dask-expr==1.0.4
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
locket==1.0.0
numpy==2.0.1
packaging==24.0
pandas==2.2.1
partd==1.4.1
pyenv-win==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
toolz==0.12.1
tzdata==2024.1
urllib3==2.2.2
yt-dlp==2024.8.1
